support: add 'flange' frame.
Similar to 'tool0', but without the orientation change (to make Z+
parallel to X+).

The 'flange' frame is intended as attachment point for EEFs, while 'tool0'
is meant to be a immutable frame that always encodes an "all-zeros tool
centre point" configuration.

Both 'tool0' and 'flange' should not be changed.